perf(codex): move identity-confuse JSON splice into helps
Keep executor files limited to executors and tests. The one-pass
string-field rewrite belongs with the other payload mutation helpers.
