+++ internal/runtime/executor/helps/payload_mutations.go
@@ -1,6 +1,9 @@
 package helps
 
 import (
+	"encoding/json"
+	"sort"
+
 	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
 )
@@ -78,4 +81,93 @@ func JoinRawJSONStrings(items []string) []byte {
 		out = append(out, item...)
 	}
 	return append(out, ']')
+}
+
+// JSONStringField is a JSON string path to overwrite in one document rewrite.
+type JSONStringField struct {
+	Path  string
+	Value string
+}
+
+// ReplaceJSONStringFields copies payload once and writes all existing string
+// fields in a single splice. Missing paths fall back to sjson.SetBytes.
+func ReplaceJSONStringFields(payload []byte, fields []JSONStringField) []byte {
+	if len(payload) == 0 || len(fields) == 0 {
+		return payload
+	}
+
+	type replacement struct {
+		index int
+		oldN  int
+		neu   []byte
+	}
+	replacements := make([]replacement, 0, len(fields))
+	var missing []JSONStringField
+	extra := 0
+	for _, field := range fields {
+		if field.Path == "" {
+			continue
+		}
+		current := gjson.GetBytes(payload, field.Path)
+		if !current.Exists() || current.Index <= 0 {
+			missing = append(missing, field)
+			continue
+		}
+		encoded, errMarshal := json.Marshal(field.Value)
+		if errMarshal != nil {
+			missing = append(missing, field)
+			continue
+		}
+		if current.Raw == string(encoded) {
+			continue
+		}
+		replacements = append(replacements, replacement{
+			index: current.Index,
+			oldN:  len(current.Raw),
+			neu:   encoded,
+		})
+		extra += len(encoded) - len(current.Raw)
+	}
+	if extra < 0 {
+		extra = 0
+	}
+
+	out := payload
+	if len(replacements) > 0 {
+		sort.Slice(replacements, func(i, j int) bool {
+			return replacements[i].index < replacements[j].index
+		})
+		buf := make([]byte, 0, len(payload)+extra)
+		last := 0
+		for _, item := range replacements {
+			if item.index < last || item.index+item.oldN > len(payload) {
+				return fallbackSetJSONStringFields(payload, fields)
+			}
+			buf = append(buf, payload[last:item.index]...)
+			buf = append(buf, item.neu...)
+			last = item.index + item.oldN
+		}
+		buf = append(buf, payload[last:]...)
+		out = buf
+	}
+	for _, field := range missing {
+		updated, errSet := sjson.SetBytes(out, field.Path, field.Value)
+		if errSet != nil {
+			continue
+		}
+		out = updated
+	}
+	return out
+}
+
+func fallbackSetJSONStringFields(payload []byte, fields []JSONStringField) []byte {
+	out := payload
+	for _, field := range fields {
+		updated, errSet := sjson.SetBytes(out, field.Path, field.Value)
+		if errSet != nil {
+			continue
+		}
+		out = updated
+	}
+	return out
 }
